backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
numpy
pydantic
motor
pymongo
youtube-transcript-api
faiss-cpu
redis
python-dotenv
google-genai
